separated simple fft spectrogram from gammatone ones
